src/main/java/org/example/ScoreCollection.java
package org.example;

import java.util.ArrayList;

public class ScoreCollection {

    private ArrayList<Integer> scores;

    public ScoreCollection(){
        this.scores = new ArrayList<Integer>();
    }

    public void addScore(Integer score) {
        this.scores.add(score);
    }

    public double averageScore(){
        double total = 0.0;

        for (int i=0; i<this.scores.size(); i++) {
            total += this.scores.get(i);
        }

        if (total == 0.0){
            return 0.0;
        }
        return total / this.scores.size();
    }
}
